+++ modules/client/CHANGELOG.md
@@ -15,8 +15,14 @@ TEMPLATE:
 ## [UPCOMING]
 ### Fixed
 - Error handling if ENS cannot be resolved.
+## [1.4.1]
+## Fixed
+- Amending the type in subgraph's `minApprovals`
+## [1.4.0]
 ### Added
 - Add ipfs optimization.
+### Fixed
+- Fix plugin installation
 ## [1.3.1]
 ### Fixed
 - Polygon network names